Add benchmark for SuppressInstrumentation
